Fixes appnavigation and loginprocessor
