+++ app/src/main/java/com/example/jetpackcomposeinstagram/navigator/AppNavigation.kt
@@ -6,16 +6,14 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.jetpackcomposeinstagram.ui.listfruititems.ListFruitItemsScreen
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 
-@OptIn(ExperimentalCoroutinesApi::class)
 @Composable
 fun AppNavigation(){
     val navController = rememberNavController()
-    NavHost(navController = navController, startDestination = Navgraphroutes.ListFruitItemsScreen.routes)
+    NavHost(navController = navController, startDestination = routes.ListFruitItemsScreen.routes)
     {
-        composable(route = Navgraphroutes.ListFruitItemsScreen.routes){
-            //ListFruitItemsScreen(modifier = Modifier)
+        composable(route = routes.ListFruitItemsScreen.routes){
+            ListFruitItemsScreen(navController, modifier = Modifier)
         }
     }
 }
