+++ app/src/main/java/com/example/jetpackcomposeinstagram/navigator/AppNavigationRoutes.kt
@@ -1,0 +1,5 @@
+package com.example.jetpackcomposeinstagram.navigator
+
+sealed class routes(var routes: String){
+    object ListFruitItemsScreen: routes("listfruititemsscreen")
+}
